+++ WatsonTcp/WatsonTcpServer.cs
@@ -501,7 +501,7 @@ namespace WatsonTcp
                         if (bytesRead > 1)
                         {
                             // check if end of headers reached
-                            if ((int)headerBuffer[0] == 58)
+                            if (headerBuffer[0] == 58)
                             {
                                 break;
                             }
@@ -726,7 +726,7 @@ namespace WatsonTcp
                     if (bytesRead > 1)
                     {
                         // check if end of headers reached
-                        if ((int)headerBuffer[0] == 58)
+                        if (headerBuffer[0] == 58)
                         {
                             break;
                         }
