+++ WatsonTcp/WatsonTcpSslClient.cs
@@ -421,7 +421,7 @@ namespace WatsonTcp
                             if (bytesRead > 1)
                             {
                                 // check if end of headers reached
-                                if ((int)headerBuffer[0] == 58)
+                                if (headerBuffer[0] == 58)
                                 {
                                     break;
                                 }
@@ -639,7 +639,7 @@ namespace WatsonTcp
                             if (bytesRead > 1)
                             {
                                 // check if end of headers reached
-                                if ((int)headerBuffer[0] == 58)
+                                if (headerBuffer[0] == 58)
                                 {
                                     break;
                                 }
