+++ WatsonTcp/WatsonTcpSslServer.cs
@@ -599,7 +599,7 @@ namespace WatsonTcp
                         if (bytesRead > 1)
                         {
                             // check if end of headers reached
-                            if ((int)headerBuffer[0] == 58)
+                            if (headerBuffer[0] == 58)
                             {
                                 break;
                             }
@@ -819,7 +819,7 @@ namespace WatsonTcp
                     if (bytesRead > 1)
                     {
                         // check if end of headers reached
-                        if ((int)headerBuffer[0] == 58)
+                        if (headerBuffer[0] == 58)
                         {
                             break;
                         }
